requirements.txt
hydra-core<=1.3.2
numpy
omegaconf<=2.3.0
pydub<=0.25.1
sentencepiece<=0.2.0
torch<=2.5.1
torchaudio<=2.5.1
onnx==1.17.0
onnxruntime==1.17.3
tqdm